Clean up active sessions list
